Added death screen
Ver. 0.9 Alpha

+++ My home programmer/Assets/Scripts/Data/SaveData.cs
@@ -4,9 +4,9 @@ public class SaveData
     public int money;
     public int Hour;
     public int Day;
-    public int EatPoints;
-    public int HealthPoints;
-    public int SleepPoints;
+    public float EatPoints;
+    public float HealthPoints;
+    public float SleepPoints;
     public int hardnes;
 }
